// packages/spreadsheet-component/src/types/commands.ts
export interface HeadersDependentCommand {
  sheetId: UID;
  dimension: Dimension;
  elements: HeaderIndex[];
}

export function isHeadersDependant(cmd: CoreCommand): boolean {
  return "dimension" in cmd && "sheetId" in cmd && "elements" in cmd;
}

export interface TargetDependentCommand {
  sheetId: UID;
  target: Zone[];
}

export function isTargetDependent(cmd: CoreCommand): boolean {
  return "target" in cmd && "sheetId" in cmd;
}

export interface RangesDependentCommand {
  ranges: RangeData[];
}

export function isRangeDependant(cmd: CoreCommand): boolean {
  return "ranges" in cmd;
}

export interface PositionDependentCommand {
  sheetId: UID;
  col: number;
  row: number;
}

export interface ZoneDependentCommand {
  sheetId: UID;
  zone: Zone;
}